Implemented sliding buffer window for increased performance
The BufferedStream will now keep up to a full chunk of consumed data at
the start of the buffer. If the consumed data grows past one chunk in
size, the unconsumed data is moved to the beginning of the buffer.

This prevents small consumes causing O(n) behavior as the chunk size
grows.

In order to make this work, a helper class for TEncoding is introduced
to allow conversion from a PByte instead of a TBytes.